# api/openapi.yaml
openapi: 3.0.3
info:
  title: Customer Self-Service Portal API
  version: "1.0.0-phase1"
  description: >
    Phase 1 (Foundation) surface: authentication, portal user management, RBAC,
    company selection, Odoo connection administration and audit log read access.
    Sales/Finance/Delivery/After-Sales/Helpdesk endpoints (spec section 11) ship
    in their own phase alongside the Odoo services that back them.


    Cross-cutting error contract for every customer-facing endpoint that reads
    Odoo (Dashboard, Sales, Invoices, Delivery, Products, Helpdesk, AfterSales,
    Equipment, Subscriptions, Analytics, Documents, Users): resolving the caller's
    Odoo context can fail in three ways that are not specific to the endpoint --
    `400 no_company_selected` (no company on the session), `403 no_identity_mapping`
    (the account has no Odoo identity on that company's connection), and
    `503 odoo_connection_unavailable` (the stored connection itself is failing --
    credential rejected or Odoo unreachable). Endpoints backed by an optional Odoo
    module add a fourth: `503 feature_unavailable`, when that module is simply not
    installed on the connected Odoo (see GET /capabilities to discover this up front
    rather than by failing). `odoo_connection_unavailable` is deliberately NOT the
    `422 odoo_auth_failed` that the admin connection endpoints return: those speak
    to an admin who just typed a credential, this one to a customer who can only be
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /rma:
    get:
      operationId: listRma
      tags: [AfterSales]
      summary: List the current user's RMA requests
      description: >
        No dedicated Odoo model exists for RMA (section 27 needs a custom addon). Each request
        is stored here and mirrored as a helpdesk.ticket; `status` is always read live from that
        ticket's stage, never cached, so it reflects whatever staff actually did in Odoo.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/RmaRequest" } }
    post:
      operationId: createRma
      tags: [AfterSales]
      summary: Submit an RMA request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason, requested_action]
              properties:
                order_id: { type: integer, description: "sale.order id this concerns, if any" }
                reason: { type: string, maxLength: 6000, description: "May contain a small sanitized HTML subset -- see src/utils/sanitizeHtml.js" }
                requested_action: { type: string, enum: [refund, replacement] }
      responses:
        "201":
          description: Created (also creates a linked helpdesk.ticket)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RmaRequest" }
...
  /equipment/due-replacements:
    get:
      operationId: getDueReplacements
      tags: [Equipment]
      summary: Fleet-wide due/overdue service parts (Fase 2, CR section 9)
      description: >
        Deterministic recommendation engine, not ML -- every row carries `basis` and `attribution`
        (IB-4), because a part bought once for a customer with several identical units can't be
        traced to one specific unit (D-5). Rate-limited per user (section 15.2, in-memory,
        per-process -- not accurate if the backend is scaled to multiple instances).
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/DueReplacementRow" } }
        "429":
          description: Rate limited
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: Installed base isn't set up on this Odoo connection (IB-5/section 15.3)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: invalid_credentials }
            message: { type: string }
            details: { type: object, nullable: true }
...
    DueReplacementRow:
      type: object
      description: >
        One deterministic recommendation (CR section 9). basis/attribution/attribution_note are
        never omitted (IB-4) -- a due date without a stated origin is never shown.
      properties:
        equipment_id: { type: integer }
        equipment_name: { type: string }
        part:
          type: object
          properties:
            product_id: { type: integer }
            name: { type: string }
        status: { type: string, enum: [overdue, due_soon, ok] }
        due_date: { type: string, format: date }
        days_overdue: { type: integer, nullable: true, description: "Only set when status=overdue" }
        basis: { type: string, enum: [last_purchase, install_date] }
        basis_date: { type: string, format: date }
        attribution:
          type: string
          enum: [unit, fleet_estimated]
          description: "'fleet_estimated' when the purchase can't be traced to one of several identical units (D-5)"
        attribution_note: { type: string, nullable: true }
        upgrade:
          type: object
          nullable: true
          description: Supersession path (CR benefit #2) -- present when this part has a recommended replacement
          properties:
            product_id: { type: integer }
            name: { type: string }
            note: { type: string, nullable: true }
...
    RmaRequest:
      type: object
      properties:
        id: { type: string, format: uuid }
        odoo_order_id: { type: integer, nullable: true }
        reason: { type: string }
        requested_action: { type: string, enum: [refund, replacement] }
        ticket_id: { type: integer, description: Linked helpdesk.ticket id }
        status:
          type: string
          nullable: true
          description: The linked ticket's current stage name, read live from Odoo
        created_at: { type: string, format: date-time }
...